refactor: move graphs to a separate package in module
