+++ app/src/main/kotlin/model/graphs/WeightedUndirectedGraph.kt
@@ -1,7 +1,7 @@
-package model
+package model.graphs
 
-import model.abstractGraph.Edge
-import model.abstractGraph.Vertex
+import model.graphs.abstractGraph.Edge
+import model.graphs.abstractGraph.Vertex
 import java.util.*
 import kotlin.NoSuchElementException
