+++ app/src/main/kotlin/model/graphs/abstractGraph/Edge.kt
@@ -1,4 +1,4 @@
-package model.abstractGraph
+package model.graphs.abstractGraph
 
 class Edge<D>(val vertex1: Vertex<D>, val vertex2: Vertex<D>) {
     fun isIncident(vertex: Vertex<D>) = (vertex1 == vertex || vertex2 == vertex)
