+++ app/src/main/kotlin/model/graphs/abstractGraph/Vertex.kt
@@ -1,3 +1,3 @@
-package model.abstractGraph
+package model.graphs.abstractGraph
 
 class Vertex<D>(var id: Int, val data: D)
