+++ app/src/main/kotlin/view/utils/SelectInitDialogWindow.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
-import model.abstractGraph.Graph
+import model.graphs.abstractGraph.Graph
 import view.tabScreen.FileControlTab
 import viewmodel.graph.CreateGraphViewModel
