+++ app/src/main/kotlin/viewmodel/graph/CreateGraphViewModel.kt
@@ -1,10 +1,10 @@
 package viewmodel.graph
 
 import androidx.compose.runtime.Composable
-import model.DirectedGraph
-import model.UndirectedGraph
-import model.WeightedDirectedGraph
-import model.WeightedUndirectedGraph
+import model.graphs.DirectedGraph
+import model.graphs.UndirectedGraph
+import model.graphs.WeightedDirectedGraph
+import model.graphs.WeightedUndirectedGraph
 import view.MainScreen
 import viewmodel.MainScreenViewModel
 
@@ -36,11 +36,14 @@ class CreateGraphViewModel {
                 when (graphStructure) {
                     is GraphStructure.Directed -> {
                         when (storedData) {
-                            is GraphType.Integer -> MainScreen(MainScreenViewModel(WeightedDirectedGraph<Int>(),
+                            is GraphType.Integer -> MainScreen(MainScreenViewModel(
+                                WeightedDirectedGraph<Int>(),
                                 "WeightedDirectedGraph Int"))
-                            is GraphType.UInteger -> MainScreen(MainScreenViewModel(WeightedDirectedGraph<UInt>(),
+                            is GraphType.UInteger -> MainScreen(MainScreenViewModel(
+                                WeightedDirectedGraph<UInt>(),
                                 "WeightedDirectedGraph UInt"))
-                            is GraphType.String -> MainScreen(MainScreenViewModel(WeightedDirectedGraph<String>(),
+                            is GraphType.String -> MainScreen(MainScreenViewModel(
+                                WeightedDirectedGraph<String>(),
                                 "WeightedDirectedGraph String"))
 
                         }
@@ -48,11 +51,14 @@ class CreateGraphViewModel {
 
                     is GraphStructure.Undirected -> {
                         when (storedData) {
-                            is GraphType.Integer -> MainScreen(MainScreenViewModel(WeightedUndirectedGraph<Int>(),
+                            is GraphType.Integer -> MainScreen(MainScreenViewModel(
+                                WeightedUndirectedGraph<Int>(),
                                 "WeightedUndirectedGraph Int"))
-                            is GraphType.UInteger -> MainScreen(MainScreenViewModel(WeightedUndirectedGraph<UInt>(),
+                            is GraphType.UInteger -> MainScreen(MainScreenViewModel(
+                                WeightedUndirectedGraph<UInt>(),
                                 "WeightedUndirectedGraph UInt"))
-                            is GraphType.String -> MainScreen(MainScreenViewModel(WeightedUndirectedGraph<String>(),
+                            is GraphType.String -> MainScreen(MainScreenViewModel(
+                                WeightedUndirectedGraph<String>(),
                                 "WeightedUndirectedGraph String")
                             )
                         }
@@ -64,22 +70,28 @@ class CreateGraphViewModel {
                 when (graphStructure) {
                     is GraphStructure.Directed -> {
                         when (storedData) {
-                            is GraphType.Integer -> MainScreen(MainScreenViewModel(DirectedGraph<Int>(),
+                            is GraphType.Integer -> MainScreen(MainScreenViewModel(
+                                DirectedGraph<Int>(),
                                 "DirectedGraph Int"))
-                            is GraphType.UInteger -> MainScreen(MainScreenViewModel(DirectedGraph<UInt>(),
+                            is GraphType.UInteger -> MainScreen(MainScreenViewModel(
+                                DirectedGraph<UInt>(),
                                 "DirectedGraph UInt"))
-                            is GraphType.String -> MainScreen(MainScreenViewModel(DirectedGraph<String>(),
+                            is GraphType.String -> MainScreen(MainScreenViewModel(
+                                DirectedGraph<String>(),
                                 "DirectedGraph String"))
                         }
                     }
 
                     is GraphStructure.Undirected -> {
                         when (storedData) {
-                            is GraphType.Integer -> MainScreen(MainScreenViewModel(UndirectedGraph<Int>(),
+                            is GraphType.Integer -> MainScreen(MainScreenViewModel(
+                                UndirectedGraph<Int>(),
                                 "UndirectedGraph Int"))
-                            is GraphType.UInteger -> MainScreen(MainScreenViewModel(UndirectedGraph<UInt>(),
+                            is GraphType.UInteger -> MainScreen(MainScreenViewModel(
+                                UndirectedGraph<UInt>(),
                                 "UndirectedGraph UInt"))
-                            is GraphType.String -> MainScreen(MainScreenViewModel(UndirectedGraph<String>(),
+                            is GraphType.String -> MainScreen(MainScreenViewModel(
+                                UndirectedGraph<String>(),
                                 "UndirectedGraph String"))
                         }
                     }
