+++ app/src/test/kotlin/model/WeightedUndirectedGraphTest.kt
@@ -1,7 +1,8 @@
 package model
 
-import model.abstractGraph.Edge
-import model.abstractGraph.Vertex
+import model.graphs.abstractGraph.Edge
+import model.graphs.abstractGraph.Vertex
+import model.graphs.WeightedUndirectedGraph
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Disabled
